refactor: split out the draw() function used by the master table
because it worked way too different compared to overlay tables. Also, in OOP it is a anti-pattern for the
parent class to know about the child classes.
